Refactor Python domain and monkey patching to prevent issues in unit tests
This commit separates all of the Python domain fixes into separate
files, as was done for C++.

Additionally, this commit ensures that any monkey patching is done
during module initialization rather than during extension setup, to
avoid applying the same monkey patch more than once if there is more
than one build performed in the same Python process, as happens during
unit tests.
